BookStore Added

+++ src/main/java/com/bookStore/repository/WalletRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bookstore.Model.Wallet;
 
-public interface WalletRepository extends JpaRepository<Wallet,Double>{
+public interface WalletRepository extends JpaRepository<Wallet,Integer>{
 
 }
